+++ Changes.md
@@ -11,6 +11,7 @@ Breaking Changes
 
 - StandardNodule : Removed deprecated `setCompatibleLabelsVisible()`.
 - DeleteAttributes : Changed base class and marked as `final`.
+- Wrapper : The `gaffer` wrapper on Linux no longer allows a custom Python build to be used. If you wish to use a custom Python, call `python _gaffer.py` instead.
 
 1.5.x.x (relative to 1.5.8.0)
 =======